restructure the service
