+++ delta.js
@@ -1,0 +1,7 @@
+import express from "express";
+
+const router = express.Router();
+
+router.post("/delta", (req, res) => {
+  console.log("receiving delta");
+});
